Ensure sidebar annots lookup works in content script
- regressed to only working in overview, where page URL is supplied
- now properly pulls the URL from the tab again if content script
