+++ src/direct-linking/background/index.ts
@@ -8,6 +8,7 @@ import { AnnotationRequests } from './request'
 import AnnotationStorage from './storage'
 import normalize from '../../util/encode-url-for-id'
 import { AnnotationSender, AnnotListEntry } from '../types'
+import { AnnotSearchParams } from 'src/search/background/types'
 
 interface TabArg {
     tab: Tabs.Tab
@@ -47,6 +48,7 @@ export default class DirectLinkingBackground {
         makeRemotelyCallable(
             {
                 createDirectLink: this.createDirectLink.bind(this),
+                getAllAnnotations: this.getAllAnnotationsByUrl.bind(this),
                 createAnnotation: this.createAnnotation.bind(this),
                 editAnnotation: this.editAnnotation.bind(this),
                 editAnnotationTags: this.editAnnotationTags.bind(this),
@@ -115,6 +117,33 @@ export default class DirectLinkingBackground {
         this.annotationStorage.indexPageFromTab(tab)
 
         return result
+    }
+
+    async getAllAnnotationsByUrl(
+        { tab }: TabArg,
+        { url, limit = 10, skip = 0, ...params }: AnnotSearchParams,
+    ) {
+        url = url == null && tab != null ? tab.url : url
+
+        const annotations = await this.annotationStorage.getAllAnnotationsByUrl(
+            {
+                url: normalize(url),
+                limit,
+                skip,
+                ...params,
+            },
+        )
+
+        return annotations.map(
+            ({ createdWhen, lastEdited, ...annotation }) => ({
+                ...annotation,
+                createdWhen: createdWhen.getTime(),
+                lastEdited:
+                    lastEdited && lastEdited instanceof Date
+                        ? lastEdited.getTime()
+                        : undefined,
+            }),
+        )
     }
 
     async createAnnotation(
