+++ src/sidebar-common/annotations-manager.ts
@@ -60,7 +60,7 @@ export default class AnnotationsManager {
         const annotationsWithoutTags: Omit<
             Annotation,
             'tags'
-        >[] = await this._getAllAnnotationsByUrlRPC(url, limit, skip)
+        >[] = await this._getAllAnnotationsByUrlRPC({ url, limit, skip })
 
         return Promise.all(
             annotationsWithoutTags.map(async annotation => {
